feat: Add support for using imgH in MedicalDatasetSR
The code changes in `dataset_medical.py` introduce a new option `use_imgH` in the `MedicalDatasetSR` class. This option allows the user to specify whether to use the `img_H` variable in the dataset. If `use_imgH` is set to `True`, the `img_H` variable is converted to a `torch.Tensor` and included in the dataset. Otherwise, it is excluded.

This change is accompanied by an update in the `test_tmi_local.json` file, where the `use_imgH` option is set to `true`. Additionally, the `H_size` option in the `train_f_u_moe_gan_local.json` file is modified to `128`.

Recent user commits and repository commits are not relevant to this specific code change.