feat: Add EDSR network implementation
The EDSR network implementation is added to the `select_network.py` file in the `models` directory. This network is used for image super-resolution and includes parameters such as the number of input and output channels, the number of features, the number of blocks, the upscale factor, and the image range. This addition enhances the functionality of the codebase.